[Refactor] Renaming the files and cleaned the project
