+++ lib/auth/landing_page/progress_dots.dart
